change: Modulariza scripts
Cria classes para os scripts de workflow.py e agent.py
